Update IMI for GEOS-Chem 14.0.0
In GEOS-Chem 14.0.0, we replaced input.geos with geoschem_config.yml.
The lines in setup_imi.sh modifying input.geos have been updated
accordingly. In addition, restart files are now stored in the Restarts
directory within the run directory. Those file paths have also been
updarted in setup_imi.sh.

Other minor fixes and updates in this commit include:
- Increase memory requested in sample ch4_template.run script.
- Do not create symbolic link to TROPOMI data if it already exists.

+++ run_imi.sh
@@ -98,7 +98,9 @@ if "$isAWS"; then
     printf "\nFinished TROPOMI download\n"
 else
     # use existing tropomi data and create a symlink to it
-    ln -s $DataPathTROPOMI $tropomiCache
+    if [[ ! -L $tropomiCache ]]; then
+	ln -s $DataPathTROPOMI $tropomiCache
+    fi
 fi
 
 ##=======================================================================
